Adjust naming and subclassing
